src/requirements.txt
Flask
flask_cors
mysql-connector